computeAggregateCovariateAnalyses {Characterization}	R Documentation

Compute aggregate covariate study

Description

Compute aggregate covariate study

Usage

computeAggregateCovariateAnalyses(
  connectionDetails = NULL,
  cdmDatabaseSchema,
  cdmVersion = 5,
  targetDatabaseSchema,
  targetTable,
  outcomeDatabaseSchema = targetDatabaseSchema,
  outcomeTable = targetTable,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  aggregateCovariateSettings,
  databaseId = "database 1",
  runId = 1
)

Arguments

connectionDetails
An object of type 'connectionDetails' as created using the [DatabaseConnector::createConnectionDetails()] function.

cdmDatabaseSchema
The schema with the OMOP CDM data

cdmVersion
The version of the OMOP CDM

targetDatabaseSchema
Schema name where your target cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

targetTable
Name of the target cohort table.

outcomeDatabaseSchema
Schema name where your outcome cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

outcomeTable
Name of the outcome cohort table.

tempEmulationSchema
Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created

aggregateCovariateSettings
The settings for the AggregateCovariate study

databaseId
Unique identifier for the database (string)

runId
Unique identifier for the tar and covariate setting

Value

The descriptive results for each target cohort in the settings.

[Package Characterization version 0.1.1 Index]