computeDechallengeRechallengeAnalyses {Characterization}	R Documentation

Compute dechallenge rechallenge study

Description

Compute dechallenge rechallenge study

Usage

computeDechallengeRechallengeAnalyses(
  connectionDetails = NULL,
  targetDatabaseSchema,
  targetTable,
  outcomeDatabaseSchema = targetDatabaseSchema,
  outcomeTable = targetTable,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  dechallengeRechallengeSettings,
  databaseId = "database 1"
)

Arguments

connectionDetails
An object of type 'connectionDetails' as created using the [DatabaseConnector::createConnectionDetails()] function.

targetDatabaseSchema
Schema name where your target cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

targetTable
Name of the target cohort table.

outcomeDatabaseSchema
Schema name where your outcome cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

outcomeTable
Name of the outcome cohort table.

tempEmulationSchema
Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created

dechallengeRechallengeSettings
The settings for the timeToEvent study

databaseId
An identifier for the database (string)

Value

An Andromeda::andromeda() object containing the dechallenge rechallenge results

[Package Characterization version 0.1.1 Index]