Compute fine the subjects that fail the dechallenge rechallenge study

Compute fine the subjects that fail the dechallenge rechallenge study

Usage

computeRechallengeFailCaseSeriesAnalyses(
  connectionDetails = NULL,
  targetDatabaseSchema,
  targetTable,
  outcomeDatabaseSchema = targetDatabaseSchema,
  outcomeTable = targetTable,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  settings,
  databaseId = "database 1",
  showSubjectId = FALSE,
  outputFolder,
  minCellCount = 0,
  ...
)

Arguments

connectionDetails

An object of type `connectionDetails` as created using the [DatabaseConnector::createConnectionDetails()] function.

targetDatabaseSchema

Schema name where your target cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

targetTable

Name of the target cohort table.

outcomeDatabaseSchema

Schema name where your outcome cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

outcomeTable

Name of the outcome cohort table.

tempEmulationSchema

Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created

settings

The settings for the timeToEvent study

databaseId

An identifier for the database (string)

showSubjectId

if F then subject_ids are hidden (recommended if sharing results)

outputFolder

A directory to save the results as csv files

minCellCount

The minimum cell value to display, values less than this will be replaced by -1

...

extra inputs

Value

An Andromeda::andromeda() object with the case series details of the failed rechallenge

See also

Examples

conDet <- exampleOmopConnectionDetails()

drSet <- createDechallengeRechallengeSettings(
  targetIds = c(1,2),
  outcomeIds = 3
)

computeRechallengeFailCaseSeriesAnalyses(
  connectionDetails = conDet,
  targetDatabaseSchema = 'main',
  targetTable = 'cohort',
  settings = drSet,
  outputFolder = tempdir()
)
#> Inputs checked
#> Connecting using SQLite driver
#> Computing dechallenge rechallenge fails results
#>   |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  11%  |                                                                              |================                                                      |  22%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================                                       |  44%  |                                                                              |=======================================                               |  56%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================                |  78%  |                                                                              |==============================================================        |  89%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.0149 secs
#> Computing dechallenge failed case series for 2 target IDs and 1 outcome IDs took 0.123 secs
#> exporting to csv file
#> Writing 0 rows to csv