exportDatabaseToCsv {Characterization}	R Documentation

Exports all tables in the result database to csv files

Description

This function extracts the database tables into csv files

Usage

exportDatabaseToCsv(
  connectionDetails,
  resultSchema,
  targetDialect,
  tablePrefix = "c_",
  filePrefix = NULL,
  tempEmulationSchema = NULL,
  saveDirectory
)

Arguments

connectionDetails
The connection details to input into the function connect in the DatabaseConnector package.

resultSchema
The name of the database schema that the result tables will be created.

targetDialect
The database management system being used

tablePrefix
The table prefix to apply to the characterization result tables

filePrefix
The prefix to apply to the files

tempEmulationSchema
The temp schema used when the database management system is oracle

saveDirectory
The directory to save the csv results

Details

This function extracts the database tables into csv files

Value

csv file per table into the saveDirectory

[Package Characterization version 0.1.1 Index]