loadCharacterizationSettings {Characterization}	R Documentation

Load the characterization settings previously saved as a json file

Description

This function converts the json file back into an R object

Usage

loadCharacterizationSettings(fileName)

Arguments

fileName
The location of the the json settings

Details

Input the directory containing the 'characterizationSettings.json' file and load the settings into R

Value

Returns the json settings as an R object

[Package Characterization version 0.1.1 Index]