saveCharacterizationSettings {Characterization}	R Documentation

Save the characterization settings as a json

Description

This function converts the settings into a json object and saves it

Usage

saveCharacterizationSettings(settings, fileName)

Arguments

settings
An object of class characterizationSettings created using createCharacterizationSettings

fileName
The location to save the json settings

Details

Input the characterization settings and output a json file to a file named 'characterizationSettings.json' inside the saveDirectory

Value

Returns the location of the drectory containing the json settings

[Package Characterization version 0.1.1 Index]