Create aggregate covariate study settings

Create aggregate covariate study settings

createAggregateCovariateSettings(
  targetIds,
  outcomeIds,
  minPriorObservation = 0,
  outcomeWashoutDays = 0,
  riskWindowStart = 1,
  startAnchor = "cohort start",
  riskWindowEnd = 365,
  endAnchor = "cohort start",
  covariateSettings = FeatureExtraction::createCovariateSettings(useDemographicsGender =
    T, useDemographicsAge = T, useDemographicsAgeGroup = T, useDemographicsRace = T,
    useDemographicsEthnicity = T, useDemographicsTimeInCohort = T,
    useDemographicsPriorObservationTime = T, useDemographicsPostObservationTime = T,
    useConditionGroupEraLongTerm = T, useDrugGroupEraLongTerm = T,
    useProcedureOccurrenceLongTerm = T, useMeasurementLongTerm = T,
    useObservationLongTerm = T, useDeviceExposureLongTerm = T,
    useVisitConceptCountLongTerm = T, 
     endDays = -1, longTermStartDays = -365),
  duringCovariateSettings = createDuringCovariateSettings(useConditionGroupEraDuring = T,
    useDrugGroupEraDuring = T, useProcedureOccurrenceDuring = T, useDeviceExposureDuring
    = T, useMeasurementDuring = T, useObservationDuring = T, useVisitConceptCountDuring =
    T),
  afterCovariateSettings =
    FeatureExtraction::createCovariateSettings(useConditionGroupEraMediumTerm = T,
    useDrugGroupEraMediumTerm = T, useProcedureOccurrenceMediumTerm = T,
    useMeasurementMediumTerm = T, useObservationMediumTerm = T,
    useDeviceExposureMediumTerm = T, useVisitConceptCountMediumTerm = T, endDays = 365,
    mediumTermStartDays = 1, ),
  minCharacterizationMean = 0
)

Arguments

targetIds

A list of cohortIds for the target cohorts

outcomeIds

A list of cohortIds for the outcome cohorts

minPriorObservation

The minimum time (in days) in the database a patient in the target cohorts must be observed prior to index

outcomeWashoutDays

Patients with the outcome within outcomeWashout days prior to index are excluded from the risk factor analysis

riskWindowStart

The start of the risk window (in days) relative to the `startAnchor`.

startAnchor

The anchor point for the start of the risk window. Can be `"cohort start"` or `"cohort end"`.

riskWindowEnd

The end of the risk window (in days) relative to the `endAnchor`.

endAnchor

The anchor point for the end of the risk window. Can be `"cohort start"` or `"cohort end"`.

covariateSettings

An object created using FeatureExtraction::createCovariateSettings

duringCovariateSettings

An object created using createDuringCovariateSettings

afterCovariateSettings

An object created using FeatureExtraction::createCovariateSettings

minCharacterizationMean

The minimum mean value for characterization output. Values below this will be cut off from output. This will help reduce the file size of the characterization output, but will remove information on covariates that have very low values. The default is 0.

Value

A list with the settings