Create aggregate covariate study settings

Create aggregate covariate study settings

Usage

createAggregateCovariateSettings(
  targetIds,
  outcomeIds,
  minPriorObservation = 0,
  outcomeWashoutDays = 0,
  riskWindowStart = 1,
  startAnchor = "cohort start",
  riskWindowEnd = 365,
  endAnchor = "cohort start",
  covariateSettings = FeatureExtraction::createCovariateSettings(useDemographicsGender =
    T, useDemographicsAge = T, useDemographicsAgeGroup = T, useDemographicsRace = T,
    useDemographicsEthnicity = T, useDemographicsIndexYear = T, useDemographicsIndexMonth
    = T, useDemographicsTimeInCohort = T, useDemographicsPriorObservationTime = T,
    useDemographicsPostObservationTime = T, useConditionGroupEraLongTerm = T,
    useDrugGroupEraOverlapping = T, useDrugGroupEraLongTerm = T,
    useProcedureOccurrenceLongTerm = T, useMeasurementLongTerm = T, 
    
    useObservationLongTerm = T, useDeviceExposureLongTerm = T,
    useVisitConceptCountLongTerm = T, useConditionGroupEraShortTerm = T,
    useDrugGroupEraShortTerm = T, useProcedureOccurrenceShortTerm = T,
    useMeasurementShortTerm = T, useObservationShortTerm = T, useDeviceExposureShortTerm
    = T, useVisitConceptCountShortTerm = T, endDays = 0, longTermStartDays = -365,
    shortTermStartDays = -30),
  caseCovariateSettings = createDuringCovariateSettings(useConditionGroupEraDuring = T,
    useDrugGroupEraDuring = T, useProcedureOccurrenceDuring = T, useDeviceExposureDuring
    = T, useMeasurementDuring = T, useObservationDuring = T, useVisitConceptCountDuring =
    T),
  casePreTargetDuration = 365,
  casePostOutcomeDuration = 365,
  extractNonCaseCovariates = T
)

Arguments

targetIds

A list of cohortIds for the target cohorts

outcomeIds

A list of cohortIds for the outcome cohorts

minPriorObservation

The minimum time (in days) in the database a patient in the target cohorts must be observed prior to index

outcomeWashoutDays

Patients with the outcome within outcomeWashout days prior to index are excluded from the risk factor analysis

riskWindowStart

The start of the risk window (in days) relative to the `startAnchor`.

startAnchor

The anchor point for the start of the risk window. Can be `"cohort start"` or `"cohort end"`.

riskWindowEnd

The end of the risk window (in days) relative to the `endAnchor`.

endAnchor

The anchor point for the end of the risk window. Can be `"cohort start"` or `"cohort end"`.

covariateSettings

An object created using FeatureExtraction::createCovariateSettings

caseCovariateSettings

An object created using createDuringCovariateSettings

casePreTargetDuration

The number of days prior to case index we use for FeatureExtraction

casePostOutcomeDuration

The number of days prior to case index we use for FeatureExtraction

extractNonCaseCovariates

Whether to extract aggregate covariates and counts for patients in the targets and outcomes in addition to the cases

Value

A list with the settings