createAggregateCovariateSettings {Characterization}	R Documentation

Create aggregate covariate study settings

Description

Create aggregate covariate study settings

Usage

createAggregateCovariateSettings(
  targetIds,
  outcomeIds,
  minPriorObservation = 0,
  riskWindowStart = 1,
  startAnchor = "cohort start",
  riskWindowEnd = 365,
  endAnchor = "cohort start",
  covariateSettings
)

Arguments

targetIds
A list of cohortIds for the target cohorts

outcomeIds
A list of cohortIds for the outcome cohorts

minPriorObservation
The minimum time in the database a patient in the target cohorts must be observed prior to index

riskWindowStart
The start of the risk window (in days) relative to the 'startAnchor'.

startAnchor
The anchor point for the start of the risk window. Can be '"cohort start"' or '"cohort end"'.

riskWindowEnd
The end of the risk window (in days) relative to the 'endAnchor'.

endAnchor
The anchor point for the end of the risk window. Can be '"cohort start"' or '"cohort end"'.

covariateSettings
An object created using FeatureExtraction::createCovariateSettings

Value

A list with the settings

[Package Characterization version 0.1.1 Index]