createCharacterizationSettings {Characterization}	R Documentation

Create the settings for a large scale characterization study

Description

This function creates a list of settings for different characterization studies

Usage

createCharacterizationSettings(
  timeToEventSettings = NULL,
  dechallengeRechallengeSettings = NULL,
  aggregateCovariateSettings = NULL
)

Arguments

timeToEventSettings
A list of timeToEvent settings

dechallengeRechallengeSettings
A list of dechallengeRechallenge settings

aggregateCovariateSettings
A list of aggregateCovariate settings

Details

Specify one or more timeToEvent, dechallengeRechallenge and aggregateCovariate settings

Value

Returns the connection to the sqlite database

[Package Characterization version 0.1.1 Index]