Create the results tables to store characterization results into a database

This function executes a large set of SQL statements to create tables that can store results

Usage

createCharacterizationTables(
  connectionDetails,
  resultSchema,
  targetDialect = "postgresql",
  deleteExistingTables = TRUE,
  createTables = TRUE,
  tablePrefix = "c_",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)

Arguments

connectionDetails

The connectionDetails to a database created by using the function createConnectDetails in the DatabaseConnector package.

resultSchema

The name of the database schema that the result tables will be created.

targetDialect

The database management system being used

deleteExistingTables

If true any existing tables matching the Characterization result tables names will be deleted

createTables

If true the Characterization result tables will be created

tablePrefix

A string appended to the Characterization result tables

tempEmulationSchema

The temp schema used when the database management system is oracle

Value

Returns NULL but creates the required tables into the specified database schema.

Details

This function can be used to create (or delete) Characterization result tables

See also

Examples

# create sqlite database
charResultDbCD <- createSqliteDatabase()

# create database results tables
createCharacterizationTables(
   connectionDetails = charResultDbCD,
   resultSchema = 'main'
 )
#> Connecting using SQLite driver
#> Deleting existing tables
#> Creating characterization results tables
#> Executing SQL took 0.00813 secs
#> Migrating data set
#> Migrator using SQL files in Characterization
#> Connecting using SQLite driver
#> Creating migrations table
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00331 secs
#> Migrations table created
#> Executing migration: Migration_1-v0_3_0_store_version.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00334 secs
#> Saving migration: Migration_1-v0_3_0_store_version.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00284 secs
#> Migration complete Migration_1-v0_3_0_store_version.sql
#> Executing migration: Migration_2-v2_1_1_mean_exposure_time_type.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  25%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00584 secs
#> Saving migration: Migration_2-v2_1_1_mean_exposure_time_type.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00464 secs
#> Migration complete Migration_2-v2_1_1_mean_exposure_time_type.sql
#> Closing database connection
#> Updating version number
#> Executing SQL took 0.00335 secs