createCharacterizationTables {Characterization}	R Documentation

Create the results tables to store characterization results into a database

Description

This function executes a large set of SQL statements to create tables that can store results

Usage

createCharacterizationTables(
  conn,
  resultSchema,
  targetDialect = "postgresql",
  deleteExistingTables = T,
  createTables = T,
  tablePrefix = "c_",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)

Arguments

conn
A connection to a database created by using the function connect in the DatabaseConnector package.

resultSchema
The name of the database schema that the result tables will be created.

targetDialect
The database management system being used

deleteExistingTables
If true any existing tables matching the Characterization result tables names will be deleted

createTables
If true the Characterization result tables will be created

tablePrefix
A string appended to the Characterization result tables

tempEmulationSchema
The temp schema used when the database management system is oracle

Details

This function can be used to create (or delete) Characterization result tables

Value

Returns NULL but creates the required tables into the specified database schema.

[Package Characterization version 0.1.1 Index]