execute a large-scale characterization study

Specify the database connection containing the CDM data, the cohort database schemas/tables, the characterization settings and the directory to save the results to

runCharacterizationAnalyses(
  connectionDetails,
  targetDatabaseSchema,
  targetTable,
  outcomeDatabaseSchema,
  outcomeTable,
  tempEmulationSchema = NULL,
  cdmDatabaseSchema,
  characterizationSettings,
  saveDirectory,
  tablePrefix = "c_",
  databaseId = "1",
  showSubjectId = F
)

Arguments

connectionDetails

The connection details to the database containing the OMOP CDM data

targetDatabaseSchema

Schema name where your target cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

targetTable

Name of the target cohort table.

outcomeDatabaseSchema

Schema name where your outcome cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

outcomeTable

Name of the outcome cohort table.

tempEmulationSchema

Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created

cdmDatabaseSchema

The schema with the OMOP CDM data

characterizationSettings

The study settings created using createCharacterizationSettings

saveDirectory

The location to save the results to

tablePrefix

A string to append the tables in the results

databaseId

The unqiue identifier for the cdm database

showSubjectId

Whether to include subjectId of failed rechallenge case series or hide

Value

An sqlite database with the results is saved into the saveDirectory and a csv file named tacker.csv details which analyses have run to completion.

Details

The results of the characterization will be saved into an sqlite database inside the specified saveDirectory