viewCharacterization - Interactively view the characterization results

This is a shiny app for viewing interactive plots and tables

Usage

viewCharacterization(resultFolder, cohortDefinitionSet = NULL)

Arguments

resultFolder

The location of the csv results

cohortDefinitionSet

The cohortDefinitionSet extracted using webAPI

Value

Opens a shiny app for interactively viewing the results

Details

Input is the output of ...

Examples

conDet <- exampleOmopConnectionDetails()

tteSet <- createTimeToEventSettings(
  targetIds = c(1,2),
  outcomeIds = 3
)

cSet <- createCharacterizationSettings(
  timeToEventSettings = tteSet
)

runCharacterizationAnalyses(
  connectionDetails = conDet,
  targetDatabaseSchema = 'main',
  targetTable = 'cohort',
  outcomeDatabaseSchema = 'main',
  outcomeTable = 'cohort',
  cdmDatabaseSchema = 'main',
  characterizationSettings = cSet,
  outputDirectory = file.path(tempdir(),'view')
)
#> Creating directory /tmp/RtmpmlJkKr/view
#> Creating directory /tmp/RtmpmlJkKr/view/execution
#> Creating new cluster
#> Connecting using SQLite driver
#> Uploading #cohort_settings
#> Inserting data took 0.00726 secs
#> Computing time to event results
#> Executing SQL took 0.024 secs
#> Computing time-to-event for 2 T-O pairs took 0.187 secs
#> exporting to csv file
#> Writing 96 rows to csv

viewCharacterization(
  resultFolder = file.path(tempdir(),'view')
)
#> Connecting using SQLite driver
#> Deleting existing tables
#> Creating characterization results tables
#> Executing SQL took 0.00742 secs
#> Migrating data set
#> Migrator using SQL files in Characterization
#> Connecting using SQLite driver
#> Creating migrations table
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00272 secs
#> Migrations table created
#> Executing migration: Migration_1-v0_3_0_store_version.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00307 secs
#> Saving migration: Migration_1-v0_3_0_store_version.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.0026 secs
#> Migration complete Migration_1-v0_3_0_store_version.sql
#> Executing migration: Migration_2-v2_1_1_mean_exposure_time_type.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |==================                                                    |  25%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================================                  |  75%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00468 secs
#> Saving migration: Migration_2-v2_1_1_mean_exposure_time_type.sql
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Executing SQL took 0.00267 secs
#> Migration complete Migration_2-v2_1_1_mean_exposure_time_type.sql
#> Closing database connection
#> Updating version number
#> Executing SQL took 0.00308 secs
#> Connecting using SQLite driver
#> Uploading file: c_time_to_event.csv to table: c_time_to_event
#> - Preparing to upload rows 1 through 96
#> Inserting data took 0.00815 secs
#> Warning: c_rechallenge_fail_case_series.csv not found
#> Warning: c_dechallenge_rechallenge.csv not found
#> Warning: c_analysis_ref.csv not found
#> Warning: c_covariate_ref.csv not found
#> Warning: c_covariates.csv not found
#> Warning: c_covariates_continuous.csv not found
#> Warning: c_settings.csv not found
#> Warning: c_cohort_details.csv not found
#> Warning: c_cohort_counts.csv not found
#> Uploading data took 0.521 secs
#> Connecting using SQLite driver
#> Inserting data took 0.00766 secs
#> Inserting data took 0.007 secs
#> Connecting using SQLite driver
#> Warning: Package OhdsiShinyModules required based on config but not installed
#> Error in utils::packageVersion(dep): there is no package called ‘OhdsiShinyModules’