viewCharacterization {Characterization}	R Documentation

viewCharacterization - Interactively view the characterization results

Description

This is a shiny app for viewing interactive plots and tables

Usage

viewCharacterization(resultLocation, cohortDefinitionSet = NULL)

Arguments

resultLocation
The location of the results

cohortDefinitionSet
The cohortDefinitionSet extracted using webAPI

Details

Input is the output of ...

Value

Opens a shiny app for interactively viewing the results

[Package Characterization version 0.1.1 Index]